Prep for 8.18.0 release
Lots of goodies here, primarily in the update to the tiered-debug module, and making the module structure here more flat to shorten names.
