+++ pyproject.toml
@@ -28,14 +28,14 @@ keywords = [
     'command-line'
 ]
 dependencies = [
-    'elasticsearch8==8.17.2',
+    'elasticsearch8==8.18.0',
     'ecs-logging==2.2.0',
     'dotmap==1.3.30',
     'click==8.1.8',
     'pyyaml==6.0.2',
     'voluptuous>=0.14.2',
     'certifi>=2025.1.31',
-    'tiered-debug==1.0.0',
+    'tiered-debug==1.1.0',
 ]
 
 [project.optional-dependencies]
