+++ src/es_client/__init__.py
@@ -1,6 +1,7 @@
 """Module Init"""
 
 from .builder import Builder
+from .debug import debug
 
-__all__ = ["Builder"]
-__version__ = "8.17.5"
+__all__ = ["Builder", "debug"]
+__version__ = "8.18.0"
