+++ src/es_client/builder.py
@@ -7,10 +7,11 @@ from dotmap import DotMap  # type: ignore
 import tiered_debug as debug
 from elastic_transport import ObjectApiResponse
 import elasticsearch8
-from es_client.helpers.schemacheck import password_filter
-from es_client.defaults import VERSION_MIN, VERSION_MAX, CLIENT_SETTINGS, OTHER_SETTINGS
-from es_client.exceptions import ConfigurationError, ESClientException, NotMaster
-from es_client.helpers.utils import (
+from .debug import debug, begin_end
+from .defaults import VERSION_MIN, VERSION_MAX, CLIENT_SETTINGS, OTHER_SETTINGS
+from .exceptions import ConfigurationError, ESClientException, NotMaster
+from .schemacheck import password_filter
+from .utils import (
     check_config,
     ensure_list,
     file_exists,
@@ -173,27 +174,25 @@ class Builder:
     def version_max(self, value) -> None:
         self.attributes.version_max = value
 
+    @begin_end()
     def set_client_defaults(self) -> None:
         """Set defaults for the client_args property"""
-        debug.lv2("Starting method...")
         self.client_args = DotMap()
         for key in CLIENT_SETTINGS:
             self.client_args[key] = None
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def set_other_defaults(self) -> None:
         """Set defaults for the other_args property"""
-        debug.lv2("Starting method...")
         self.other_args = DotMap()
         for key in OTHER_SETTINGS:
             self.other_args[key] = None
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def process_config_opts(
         self, configdict: t.Union[t.Dict, None], configfile: t.Union[str, None]
     ) -> None:
         """Process whether to use a configdict or configfile"""
-        debug.lv2("Starting method...")
         if configfile:
             debug.lv2(f'Using values from file: {configfile}')
             self.config = check_config(get_yaml(configfile))
@@ -204,11 +203,10 @@ class Builder:
             # Empty/Default config.
             debug.lv2("No configuration file or dictionary provided. Using defaults.")
             self.config = check_config({"client": {}, "other_settings": {}})
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def update_config(self) -> None:
         """Update object with values provided"""
-        debug.lv2("Starting method...")
         self.client_args.update(self.config.client)
         self.other_args.update(self.config.other_settings)
         self.master_only = self.other_args.master_only
@@ -217,11 +215,10 @@ class Builder:
             self.skip_version_test = self.other_args.skip_version_test
         else:
             self.skip_version_test = False
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def validate(self) -> None:
         """Validate that what has been supplied is acceptable to attempt a connection"""
-        debug.lv2("Starting method...")
         # Configuration pre-checks
         if self.client_args.hosts is not None:
             verified_hosts = []
@@ -242,11 +239,10 @@ class Builder:
         self._check_api_key()
         self._check_cloud_id()
         self._check_ssl()
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def connect(self) -> None:
         """Attempt connection and do post-connection checks"""
-        debug.lv2("Starting method...")
         # Get the client
         self._get_client()
         # Post checks
@@ -255,11 +251,10 @@ class Builder:
             self._check_multiple_hosts()
             self._find_master()
             self._check_if_master()
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def _check_basic_auth(self) -> None:
         """Create ``basic_auth`` tuple from username and password"""
-        debug.lv2("Starting method...")
         if "username" in self.other_args or "password" in self.other_args:
             usr = self.other_args.username if "username" in self.other_args else None
             pwd = self.other_args.password if "password" in self.other_args else None
@@ -272,8 +267,8 @@ class Builder:
                 raise ConfigurationError(msg)
             else:
                 self.client_args.basic_auth = (usr, pwd)
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def _check_api_key(self) -> None:
         """
         Create ``api_key`` tuple from :py:attr:`other_args` ``['api_key']`` subkeys
@@ -282,7 +277,6 @@ class Builder:
         Or if ``api_key`` subkey ``token`` is present, derive ``id`` and ``api_key``
         from ``token``
         """
-        debug.lv2("Starting method...")
         if "api_key" in self.other_args:
             # If present, token will override any value in 'id' or 'api_key'
             # pylint: disable=no-member
@@ -311,11 +305,10 @@ class Builder:
                     raise ConfigurationError(msg)
                 else:
                     self.client_args.api_key = (api_id, api_key)
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def _check_cloud_id(self) -> None:
         """Remove ``hosts`` key if ``cloud_id`` provided"""
-        debug.lv2("Starting method...")
         if "cloud_id" in self.client_args and self.client_args.cloud_id is not None:
             # We can remove the default if that's all there is
             if (
@@ -329,14 +322,13 @@ class Builder:
                     'ConfigurationError: "Cannot populate both hosts and cloud_id"'
                 )
                 raise ConfigurationError('Cannot populate both "hosts" and "cloud_id"')
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def _check_ssl(self) -> None:
         """
         Use `certifi <https://github.com/certifi/python-certifi>`_ if using ssl
         and ``ca_certs`` has not been specified.
         """
-        debug.lv2("Starting method...")
         verify_ssl_paths(self.client_args)
         if "cloud_id" in self.client_args and self.client_args.cloud_id is not None:
             scheme = "https"
@@ -361,19 +353,17 @@ class Builder:
                             debug.lv3('Exiting method, raising exception')
                             debug.lv5(f'Exception = "{msg}"')
                             raise ConfigurationError(msg)
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def _find_master(self) -> None:
         """Find out if we are connected to the elected master node"""
-        debug.lv2("Starting method...")
         my_node_id = list(self.client.nodes.info(node_id="_local")["nodes"])[0]
         master_node_id = self.client.cluster.state(metric="master_node")["master_node"]
         self.is_master = my_node_id == master_node_id
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def _check_multiple_hosts(self) -> None:
         """Check for multiple hosts when master_only"""
-        debug.lv2("Starting method...")
         if "hosts" in self.client_args and isinstance(self.client_args.hosts, list):
             if len(self.client_args.hosts) > 1:
                 debug.lv3('Exiting method, raising exception')
@@ -383,14 +373,13 @@ class Builder:
                 )
                 logger.error(msg)
                 raise ConfigurationError(msg)
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def _check_if_master(self) -> None:
         """
         If we are not connected to the elected master node, raise
         :py:exc:`~es_client.exceptions.NotMaster`
         """
-        debug.lv2("Starting method...")
         if not self.is_master:
             msg = (
                 "The master_only flag is set to True, but the client is  "
@@ -399,14 +388,13 @@ class Builder:
             debug.lv3('Exiting method, raising exception')
             logger.error(msg)
             raise NotMaster(msg)
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def _check_version(self) -> None:
         """
         Compare the Elasticsearch cluster version to :py:attr:`min_version` and
         :py:attr:`max_version`
         """
-        debug.lv2("Starting method...")
         v = get_version(self.client)
         if self.skip_version_test:
             logger.warning("Skipping Elasticsearch version checks")
@@ -417,26 +405,24 @@ class Builder:
                 debug.lv3('Exiting method, raising exception')
                 logger.error(msg)
                 raise ESClientException(msg)
-        debug.lv3('Exiting method')
 
+    @begin_end()
     def _get_client(self) -> None:
         """
         Instantiate the
         :py:class:`~.elasticsearch.Elasticsearch` object and populate
         :py:attr:`client`
         """
-        debug.lv2('Starting method...')
         # Eliminate any remaining "None" entries from the client arguments
         client_args = prune_nones(self.client_args.toDict())
         self.client = elasticsearch8.Elasticsearch(**client_args)
 
+    @begin_end()
     def test_connection(self) -> ObjectApiResponse[t.Any]:
         """
         Connect and execute :meth:`Elasticsearch.info()
         <elasticsearch8.Elasticsearch.info>`
         """
-        debug.lv2('Starting method...')
         retval = self.client.info()
-        debug.lv3('Exiting method, returning value')
-        debug.lv5(f'Value = "{retval}"')
+        debug.lv5(f'Return value = "{retval}"')
         return retval
