+++ tests/unit/test_helpers_utils.py
@@ -8,7 +8,7 @@ from unittest import TestCase
 from unittest.mock import Mock
 import pytest
 from es_client.exceptions import ConfigurationError
-from es_client.helpers import utils as u
+from es_client import utils as u
 from . import FileTestObj
 
 # pylint: disable=R0903,W0718
